remove unused

+++ x/tss/types/keys.go
@@ -75,5 +75,5 @@ func AccumulatedCommitStoreKey(groupID tss.GroupID) []byte {
 }
 
 func AccumulatedCommitIndexStoreKey(groupID tss.GroupID, index uint64) []byte {
-	return append(AccumulatedCommitStoreKey(groupID), sdk.Uint64ToBigEndian(uint64(index))...)
+	return append(AccumulatedCommitStoreKey(groupID), sdk.Uint64ToBigEndian(index)...)
 }
